Added EPaymentMethod enums, written UserPresentation, fixed UserService Delete function

+++ CompanySupportTicketSystem.Service/Interfaces/IUserService.cs
@@ -7,7 +7,7 @@ public interface IUserService
 {
     public Task<bool> AddAsync(User user);
     public Task<bool> UpdateAsync(long id,UserForUpdateDto dto);
-    public Task<bool> DeleteByIdAsync(int id);
-    public Task<UserForResultDto> GetByIdAsync(int id);
+    public Task<bool> DeleteByIdAsync(long id);
+    public Task<UserForResultDto> GetByIdAsync(long id);
     public Task<IEnumerable<UserForResultDto>> GetAllAsync();
 }
